Create a service for JSON
