+++ registry/server/common/services/preProcessResponse.ts
@@ -1,25 +1,7 @@
 import _ from 'lodash/fp';
-
-/**
- * Original source code was taken from {@link https://github.com/prototypejs/prototype/blob/5fddd3e/src/prototype/lang/string.js#L702}
- */
-const isJSON = (str: string): boolean => {
-    if (/^\s*$/.test(str)) return false;
-
-    str = str.replace(/\\(?:["\\\/bfnrt]|u[0-9a-fA-F]{4})/g, '@');
-    str = str.replace(/"[^"\\\n\r]*"|true|false|null|-?\d+(?:\.\d*)?(?:[eE][+\-]?\d+)?/g, ']');
-    str = str.replace(/(?:^|:|,)(?:\s*\[)+/g, '');
-
-    return (/^[\],:{}\s]*$/).test(str);
-};
-
-const parseJSON = (value: any): any => {
-    if (_.isString(value) && isJSON(value)) {
-        return JSON.parse(value);
-    }
-
-    return value;
-}
+import {
+    parseJSON,
+} from './json';
 
 const omitEmptyValues = _.omitBy(_.cond([
     [_.isNull, _.stubTrue],
@@ -35,15 +17,9 @@ const omitPreResponseEmptyValues = _.cond([
     [_.stubTrue, (value: any) => value]
 ]);
 
-const parsePreResponseJSON = _.cond([
-    [_.isArray, _.map(_.mapValues(parseJSON))],
-    [_.isObject, _.mapValues(parseJSON)],
-    [_.stubTrue, parseJSON]
-]);
-
 const preProcessResponse = _.compose(
     omitPreResponseEmptyValues,
-    parsePreResponseJSON,
+    parseJSON,
 );
 
 export default preProcessResponse;
